solved error in program

+++ lib/widget/newTransaction.dart
@@ -16,7 +16,7 @@ class _NewTransactionState extends State<NewTransaction> {
 
   void submitData() {
     final enteredTitle = titleController.text;
-    final enteredAmount = double.parse(amountController.text);
+    final enteredAmount = int.parse(amountController.text);
     if (enteredTitle.isEmpty || enteredAmount < 1) {
       return;
     }
